requirements.txt
appdirs==1.4.3
asgiref==3.3.4
crispy-bootstrap5==0.4
distlib==0.3.0
Django==3.2.4
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==1.6.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.10.0.0
virtualenv==20.0.18
zipp==3.1.0
